consolidate JetTagProducer into one file
